+++ build.rs
@@ -1,9 +1,4 @@
-extern crate winresource;
-
 fn main() {
-    if std::env::var("CARGO_CFG_TARGET_OS").unwrap() == "windows" {
-        let mut res = winresource::WindowsResource::new();
-        res.set_icon("app.ico");
-        res.compile().unwrap();
-    }
+    #[cfg(windows)]
+    embed_resource::compile("./windows/work-break.rc", embed_resource::NONE);
 }